# backend/docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact: {}
  description: 这是一个应用版本管理系统的API服务
  title: 应用版本管理系统 API
  version: "1.0"
paths:
  /open/changelog:
    get:
      consumes:
      - application/json
      description: 获取应用的版本更新历史记录
      parameters:
      - description: 应用标识
        in: query
        name: identifier
        required: true
        type: string
      - description: 平台筛选
        in: query
        name: platform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 版本更新历史列表
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "404":
          description: 应用不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取应用更新日志
      tags:
      - 开放接口
  /open/latest:
    get:
      consumes:
      - application/json
      description: 根据应用标识和平台获取最新版本信息，支持自定义输出格式
      parameters:
      - description: 应用标识
        in: query
        name: identifier
        required: true
        type: string
      - description: 平台 (android/ios/windows/macos/linux/harmony)
        in: query
        name: platform
        type: string
      - description: 输出格式，对应模板名称
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回最新版本信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 应用或版本不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取应用最新版本
      tags:
      - 开放接口
swagger: "2.0"